+++ backend/backend/src/wrapper/entity/session/dto.rs
@@ -7,8 +7,8 @@ use crate::wrapper::entity::user::User;
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, ToSchema)]
 pub(crate) struct PublicSession {
-    pub(crate) id: i32,
-    pub(crate) user_id: i32,
+    pub(crate) id: i64,
+    pub(crate) user_id: i64,
     pub(crate) name: String,
     pub(crate) description: Option<String>,
     pub(crate) platform: Option<String>,
